feat: log client connection error
If there is no “error” listener defined, and an “error” event is emitted, the error is thrown and the node process exists.

This behaviour is documented in Node.js EventEmitter

https://nodejs.org/api/events.html#events_error_events

In the context of node-postgres, this means that even if there is an active query to associate an error with, when there is no “error” listener defined the program will exit when there is a connection error.

Refer to:

https://github.com/brianc/node-postgres/issues/1700