feat: implemenet connection pooling
pg-pool have removed `min` option (https://github.com/brianc/node-pg-pool/issues/135). Our implementation partially implements the original pro-active pooling logic. However, unlike the original implementation, we only pool new queries when another query is complete. The benefit of this approach is that if the process is otherwise idle, there are no unnecessary sockets being kept. Equally, the moment a single query is issued, we ensure that we bring the pool size to the minimum number of connections expected.