feat: use custom binding syntax to avoid ambiguity when parsing statements containing references to positional paraameters
Earier, if you had a query such as sql`SELECT '$1'` Slonik would have thought that $1 is referring to a parameter that is expected to be bound to the query.
Slonik bindings produce custom placeholders (e.g. #$#1) that allow us to distinguish between values bound using Slonik and those added by the user.